Trigger fixes

+++ database/migrations/1550233065149_delete_like_action_schema.js
@@ -9,9 +9,7 @@ class DeleteLikeActionSchema extends Schema {
         await Database
             .raw(
                 'CREATE TRIGGER `delete_like_action` BEFORE DELETE ON `likes`' +
-                'FOR EACH ROW DELETE FROM actions' +
-                'WHERE receiver_id = IF(OLD.type = 1, (SELECT owner_id FROM posts WHERE id=OLD.entity_id), (SELECT owner_id FROM comments WHERE id=OLD.entity_id))' +
-                'AND initiator_id=OLD.owner_id AND type=1 AND entity_id=OLD.id'
+                'FOR EACH ROW DELETE FROM actions WHERE receiver_id = IF(OLD.type = 1, (SELECT owner_id FROM posts WHERE id=OLD.entity_id), (SELECT owner_id FROM comments WHERE id=OLD.entity_id)) AND initiator_id=OLD.owner_id AND type=1 AND entity_id=OLD.id'
             );
     }
